# docs/swagger.yaml
definitions:
  github_com_XTrau_auth-service_internal_dto.LoginRequest:
    properties:
      login:
        maxLength: 64
        minLength: 8
        type: string
      password:
        maxLength: 64
        minLength: 8
        type: string
    type: object
  github_com_XTrau_auth-service_internal_dto.RegisterRequest:
    properties:
      password:
        maxLength: 64
        minLength: 8
        type: string
      username:
        maxLength: 64
        minLength: 8
        type: string
    type: object
  github_com_XTrau_auth-service_internal_dto.UserDataResponse:
    properties:
      username:
        type: string
    type: object
  internal_handlers.AuthHandlers:
    type: object
info:
  contact: {}
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: |-
        Возвращает access token в body и refresh token в cookie
        Для доступа к api нужен access token
        Время жизни access token - 15 минут
        Время жизни refresh token - 30 дней
        Для обновления access token необходимо делать запрос POST /refresh с refresh token в куках
      parameters:
      - description: userData
        in: body
        name: userObject
        required: true
        schema:
          $ref: '#/definitions/github_com_XTrau_auth-service_internal_dto.LoginRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.AuthHandlers'
      summary: Логин пользователя
      tags:
      - Аутентификация
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Удаляет refresh token из cookie
      responses:
        "204":
          description: No Content
      summary: Разлогинить пользователя
      tags:
      - Аутентификация
  /auth/refresh:
    post:
      description: |-
        Возвращает новый access и refresh токены
        Требуется refresh токен в куках
      responses:
        "200":
          description: OK
      summary: Обновить токены
      tags:
      - Аутентификация
  /auth/register:
    post:
      consumes:
      - application/json
      description: |-
        Регистрирует пользователя в системе
        логиниться в системе нужно отдельно от регистрации
      parameters:
      - description: userData
        in: body
        name: userObject
        required: true
        schema:
          $ref: '#/definitions/github_com_XTrau_auth-service_internal_dto.RegisterRequest'
      responses:
        "201":
          description: Created
      summary: Регистрация пользователя
      tags:
      - Аутентификация
  /auth/user:
    get:
      description: Возвращает данные пользователя по access_token
      parameters:
      - description: 'Example: Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_XTrau_auth-service_internal_dto.UserDataResponse'
        "401":
          description: unauthorized user
          schema:
            type: string
      summary: Получить пользователя
      tags:
      - Аутентификация
swagger: "2.0"